requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.11
Deprecated==1.2.15
devtools==0.12.2
distro==1.9.0
executing==2.0.1
feedparser==6.0.11
filelock==3.16.1
finviz-platform @ git+https://github.com/mariostoev/finviz@bc02ca9bb980151c33fc7000276bbb0a0c729577
flatbuffers==24.3.25
fonttools==4.55.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.7
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
namex==0.0.8
narwhals==1.15.1
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.55.3
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
patsy==1.0.1
peewee==3.17.6
pillow==11.0.0
platformdirs==4.2.2
pmdarima==2.0.4
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
statsmodels==0.14.4
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.66.5
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
user_agent==0.1.10
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
win32-setctime==1.1.0
wrapt==1.17.0
yarl==1.9.4
yfinance==0.2.50
